An Android Application for the students to evaluate the performance of their teachers on the basis of several factors.
This application will provide an intiutive user interface which will help the students to asess the performance of teachers on the basis of several factors such as radability, preperations for lectures etc. in order to improve the teaching style of the teachers by evaluating on which factors the teacher needs improvement.